Call Modal

+++ client/src/components/message/RightSide.js
@@ -117,8 +117,10 @@ const RightSide = () => {
 	};
 
 	const handleDeleteConversation = () => {
-		dispatch(deleteConversation({ auth, id }));
-		return history.push('/message');
+		if (window.confirm('Do you want to delete this conversation? ')) {
+			dispatch(deleteConversation({ auth, id }));
+			return history.push('/message');
+		}
 	};
 
 	//Get message from param userId => set content chat container
@@ -165,15 +167,39 @@ const RightSide = () => {
 		// eslint-disable-next-line
 	}, [isLoadMore]);
 
+	// Call Audio And Call Video
+
+	const caller = ({ video }) => {
+		const { _id, avatar, username, fullname } = user;
+		const msg = {
+			sender: auth.user._id,
+			recipient: _id,
+			avatar,
+			username,
+			fullname,
+			video
+		};
+		dispatch({ type: GLOBALTYPES.CALL, payload: msg });
+	};
+	const handleAudioCall = () => {
+		caller({ video: false });
+	};
+	const handleVideoCall = () => {
+		caller({ video: true });
+	};
 	return (
 		<>
 			<div className='message_header' style={{ cursor: 'pointer' }}>
 				{user.length !== 0 && (
 					<UserCard user={user}>
-						<i
-							className='fas fa-trash text-danger'
-							onClick={handleDeleteConversation}
-						/>
+						<div>
+							<i className='fas fa-phone-alt' onClick={handleAudioCall} />
+							<i className='fas fa-video mx-3' onClick={handleVideoCall} />
+							<i
+								className='fas fa-trash text-danger'
+								onClick={handleDeleteConversation}
+							/>
+						</div>
 					</UserCard>
 				)}
 			</div>
